% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindCliques.R
\name{CliquePartition}
\alias{CliquePartition}
\title{Clique Partition}
\usage{
CliquePartition(clique.set, randomize = F)
}
\arguments{
\item{clique.set}{Input set of cliques which may overlap}

\item{randomize}{Whether we randomize the input list}
}
\value{
A list of non-overlapping cliques
}
\description{
Clique Partition
}
