% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateDistances.R
\name{CliquesCount}
\alias{CliquesCount}
\title{Count of Cliques}
\usage{
CliquesCount(cliques)
}
\arguments{
\item{cliques}{A list of clique indices}
}
\value{
List of indices corresponding to cliques and corresponding vector of labels
}
\description{
Count of Cliques
}
