% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindCliques.R
\name{CliqueSearch}
\alias{CliqueSearch}
\title{Exhaustive Search For Cliques}
\usage{
CliqueSearch(G, min_clique_size = 8, verbose = F)
}
\arguments{
\item{G}{Adjacency Matrix}

\item{min_clique_size}{Minimum clique size}

\item{verbose}{Whether to include additional messages}
}
\value{
A list of non-overlapping cliques
}
\description{
Exhaustive Search For Cliques
}
