% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindCliques.R
\name{ClusterCliqueSearch}
\alias{ClusterCliqueSearch}
\title{Search For Cliques}
\usage{
ClusterCliqueSearch(G, min_clique_size = 8, res = 1, verbose = F)
}
\arguments{
\item{G}{Adjacency Matrix}

\item{min_clique_size}{Minimum clique size}

\item{res}{Tuning parameter for clustering}

\item{verbose}{Whether to include additional messages}
}
\value{
A list of non-overlapping cliques
}
\description{
Search For Cliques
}
