% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsampleConstantCurvatureTest.R
\name{ComputePvalue}
\alias{ComputePvalue}
\title{Compute p-value from overlap of the subsampled upper and lower bounds.}
\usage{
ComputePvalue(upper.bounds, lower.bounds)
}
\arguments{
\item{upper.bounds}{upper bound subsample set}

\item{lower.bounds}{lower bound subsample set}
}
\value{
p-value from subsampled upper and lower bounds.
}
\description{
Compute p-value from overlap of the subsampled upper and lower bounds.
}
