% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateCurvature.R
\name{ConstantCurvatureTest}
\alias{ConstantCurvatureTest}
\title{Test for Constant curvature of a distance matrix}
\usage{
ConstantCurvatureTest(
  D,
  tri.const = 1.4,
  num.midpoints = 3,
  d.yz.min = 1.5,
  d.yz.max = Inf,
  curve.scale = 10,
  verbose = F
)
}
\arguments{
\item{D}{Estimated Distance matrix}

\item{tri.const}{Triangle Constant to use for selecting good x points}

\item{num.midpoints}{Number of midpoint sets to use}

\item{d.yz.min}{Minimum distance of y z pairs to consider}

\item{d.yz.max}{Maximum distance of y z pairs to consider}

\item{curve.scale}{Threshold to use to minimize effect of outliers}

\item{verbose}{Additional plotting details}
}
\value{
Test for constant curvature of a given distance matrix
}
\description{
Test for Constant curvature of a distance matrix
}
