% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsampleConstantCurvatureTest.R
\name{estimateBounds}
\alias{estimateBounds}
\title{estimate curvature bounds using a reference set x}
\usage{
estimateBounds(D, y, z, m, x.set)
}
\arguments{
\item{D}{Distance Matrix}

\item{y}{reference point y}

\item{z}{reference point z}

\item{m}{surrogate midpoint of y and z}

\item{x.set}{set of reference x coordinates}
}
\value{
Upper And Lower Bounds for estimate
}
\description{
estimate curvature bounds using a reference set x
}
