% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateDistances.R
\name{EstimateD}
\alias{EstimateD}
\title{Estimate Distance Matrix From Cliques}
\usage{
EstimateD(
  G,
  cliques,
  D0,
  thresh = 10^(-3),
  max.iter = 50,
  solver = "MOSEK",
  verbose = F,
  rand.eff.0 = F
)
}
\arguments{
\item{G}{Full Data Adjacency Matrix}

\item{cliques}{List of indices corresponding to cliques}

\item{D0}{Initial Point To Take Taylor Series of Objective}

\item{thresh}{Default Threshold for Increasing Likelihood}

\item{max.iter}{Maximum number of iterations of the second order approximation}

\item{solver}{Which solver should be passed to CVX}

\item{verbose}{Include additional messages}

\item{fixed.effect.vec}{List of estimated fixed Effects}
}
\value{
Estimated Distance Matrix
}
\description{
Estimate Distance Matrix From Cliques
}
