% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateDistances.R
\name{EstimateDFrob}
\alias{EstimateDFrob}
\title{Estimate Distance Matrix From Cliques (Frobenius Norm)}
\usage{
EstimateDFrob(G, cliques, solver = "MOSEK", verbose = F, rand.eff.0 = F)
}
\arguments{
\item{G}{Full Data Adjacency Matrix}

\item{cliques}{List of indices corresponding to cliques}

\item{solver}{Which solver should be passed to CVX}

\item{verbose}{Include additional messages}

\item{fixed.effect.vec}{List of estimated fixed Effects}
}
\value{
Estimated Distance Matrix
}
\description{
Estimate Distance Matrix From Cliques (Frobenius Norm)
}
