% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateCurvature.R
\name{EstimateKappa}
\alias{EstimateKappa}
\title{Find the solution to the implicit estimating function to estimate curvature}
\usage{
EstimateKappa(
  dxy,
  dxz,
  dyz,
  dxm,
  kappa.init = 0,
  thresh = 10^(-6),
  max.iter = 10,
  max.curve = Inf,
  ee.thresh = 0.1,
  kappa.reset.step = 10
)
}
\arguments{
\item{dxy}{Triangle Side Length}

\item{dxz}{Triangle Side Length}

\item{dyz}{Triangle Side Length}

\item{dxm}{Length from point x to the midpoint of y and z}

\item{kappa.init}{Initializing Curvature}

\item{thresh}{Numerical precision}

\item{max.iter}{Maximum number of iterations}

\item{max.curve}{Bound for the Curvature}

\item{ee.thresh}{Threshold for the solution of the estimating function.}

\item{kappa.reset.step}{}
}
\value{
Estimated Curvature Value
}
\description{
Find the solution to the implicit estimating function to estimate curvature
}
\examples{
EstimateKappa(1,1,1,sqrt(3/4))
}
