% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateCurvature.R
\name{EstimateKappaSet}
\alias{EstimateKappaSet}
\title{Estimate a set of curvatures using a set of indices by x}
\usage{
EstimateKappaSet(D, y, z, m, x.set)
}
\arguments{
\item{D}{Distance matrix}

\item{y}{Index of midpoint set}

\item{z}{Index of midpoint set}

\item{m}{Index of midpoint set}

\item{x.set}{Vector of indices for which we want to estimate}
}
\value{
Vector of kappa estimates corresponding to x.set indices
}
\description{
Estimate a set of curvatures using a set of indices by x
}
