% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateDistances.R
\name{estimateRandeff}
\alias{estimateRandeff}
\title{Estimate Random Effects By Degree Ratios}
\usage{
estimateRandeff(G, cliques)
}
\arguments{
\item{G}{Full Data Adjacency Matrix}

\item{cliques}{List of indices corresponding to cliques}
}
\value{
Estimated Set of Random Effects
}
\description{
Estimate Random Effects By Degree Ratios
}
