% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicMath.R
\name{expit}
\alias{expit}
\title{Expit Function}
\usage{
expit(x)
}
\arguments{
\item{x}{Input value to take expit}
}
\value{
Expit, or the logistic function of x
}
\description{
Expit Function
}
\examples{
expit(0)
}
