% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateCurvature.R
\name{filter_indices}
\alias{filter_indices}
\title{Triangle Inequality Filter For Distance Matrices}
\usage{
filter_indices(D, y, z, m, tri.const = sqrt(2))
}
\arguments{
\item{D}{Distance matrix}

\item{y}{Index of midpoint set}

\item{z}{Index of midpoint set}

\item{m}{Index of midpoint set}

\item{tri.const}{Scaling constant for triangle inequality}
}
\value{
Indices which satisfy the stronger triangle inequality
}
\description{
Triangle Inequality Filter For Distance Matrices
}
