% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateDistances.R
\name{FWA}
\alias{FWA}
\title{Floyd-Warshall Algorithm}
\usage{
FWA(D0.init)
}
\arguments{
\item{D0.init}{Initialization for Floyd-Warshall Algorithm}
}
\value{
Distance Matrix which satisfied triangle inequality
}
\description{
Floyd-Warshall Algorithm
}
