% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsampleConstantCurvatureTest.R
\name{g_u}
\alias{g_u}
\title{Estimating function for the upper bound on curvature}
\usage{
g_u(kappa, dxy, dxz, dyz, dxm, dym, dzm)
}
\arguments{
\item{kappa}{curvature}

\item{dxy}{distance between points x and y}

\item{dxz}{distance between points x and z}

\item{dyz}{distance between points y and z}

\item{dxm}{distance between points x and m}

\item{dym}{distance between points y and m}

\item{dzm}{distance between points z and m}
}
\value{
curvature estimating function bound
}
\description{
Estimating function for the upper bound on curvature
}
