% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateCurvature.R
\name{gGradKappa}
\alias{gGradKappa}
\title{Derivative of implicit estimating function}
\usage{
gGradKappa(kappa, dxy, dxz, dyz, dxm)
}
\arguments{
\item{kappa}{Curvature of Embedding Space}

\item{dxy}{Triangle Side Length}

\item{dxz}{Triangle Side Length}

\item{dyz}{Triangle Side Length}

\item{dxm}{Length from point x to the midpoint of y and z}
}
\value{
Gradient of the estimating function
}
\description{
Derivative of implicit estimating function
}
\examples{
gGradKappa(0,1,1,1,1/2)
}
