% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateDistances.R
\name{InitD0}
\alias{InitD0}
\title{Good Initialization For Distance Matrix}
\usage{
InitD0(G, cliques, fixed.effect.vec, global.randeff = T, add.offset = 0.001)
}
\arguments{
\item{G}{Full Data Adjacency Matrix}

\item{cliques}{List of indices corresponding to cliques}

\item{fixed.effect.vec}{Vector of corresponding estimated random effects}

\item{global.randeff}{Whether to approximate and group the random effects into a single one}

\item{add.offset}{Add a minimum distance for better optimization}
}
\value{
Initialization of Estimated Distance Matrix which obeys triangle inequalities
}
\description{
Good Initialization For Distance Matrix
}
