% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsampleConstantCurvatureTest.R
\name{kappa_u}
\alias{kappa_u}
\title{Estimator of upper bound on the curvature}
\usage{
kappa_u(
  dxy,
  dxz,
  dyz,
  dxm,
  dym,
  dzm,
  kappa.prec = 10^(-5),
  min.curvature = -1000
)
}
\arguments{
\item{dxy}{distance between points x and y}

\item{dxz}{distance between points x and z}

\item{dyz}{distance between points y and z}

\item{dxm}{distance between points x and m}

\item{dym}{distance between points y and m}

\item{dzm}{distance between points z and m}

\item{kappa.prec}{granularity of kappa grid}

\item{min.curvature}{minimum curvature value searched}
}
\value{
Upper Bound Estimate
}
\description{
Estimator of upper bound on the curvature
}
