% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateCurvature.R
\name{MidDist}
\alias{MidDist}
\title{Value of dxm as a function of the triangle side lengths and the curvature}
\usage{
MidDist(kappa, dxy, dxz, dyz)
}
\arguments{
\item{kappa}{Curvature of Embedding Space}

\item{dxy}{Triangle Side Length}

\item{dxz}{Triangle Side Length}

\item{dyz}{Triangle Side Length}
}
\value{
Distance of the length xm under the specified embedding space
}
\description{
Value of dxm as a function of the triangle side lengths and the curvature
}
\examples{
MidDist(0,1,1,1)
}
