% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateCurvature.R
\name{optimal_midpoint_search}
\alias{optimal_midpoint_search}
\title{Midpoint Search Function}
\usage{
optimal_midpoint_search(D, top.k = 1, d.yz.min = 1, d.yz.max = 2.5)
}
\description{
Midpoint Search Function
}
