% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicMath.R
\name{sec}
\alias{sec}
\title{Secant Function}
\usage{
sec(x)
}
\arguments{
\item{x}{Input value to take secant}
}
\value{
Secant of x, defined as the reciprocal of the cosine.
}
\description{
Secant Function
}
\examples{
sec(0)
}
