% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsampleConstantCurvatureTest.R
\name{selectReference}
\alias{selectReference}
\title{Selection of good midpoints and reference points. Simple default values are included for ease of use}
\usage{
selectReference(D, J, tri.const = 1.4, d.yz.min = 1, d.yz.max = 5, sub.idx)
}
\arguments{
\item{D}{Distance Matrix}

\item{J}{Number of midpoints}

\item{tri.const}{filter triangle constant}

\item{d.yz.min}{minimum distance for yz}

\item{d.yz.max}{maximum distance for yz}

\item{sub.idx}{subset of indices searched}
}
\value{
Reference points
}
\description{
Selection of good midpoints and reference points. Simple default values are included for ease of use
}
