% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsampleConstantCurvatureTest.R
\name{SubSampleConstantCurvatureTest}
\alias{SubSampleConstantCurvatureTest}
\title{Constant Curvature Test Based on Subsampling approaches.}
\usage{
SubSampleConstantCurvatureTest(
  A,
  clique.set,
  reference.set,
  subsample.rate = 1,
  B = 100,
  store.D.samples = T,
  max.iter = 1
)
}
\arguments{
\item{A}{Adjacency matrix}

\item{clique.set}{indices of where the cliques are located in A}

\item{reference.set}{reference set for where curvature is measured}

\item{subsample.rate}{drop out number for the subsampling distributions}

\item{B}{Number of subsample iterations}

\item{store.D.samples}{stores subsamples of the matrix D for later use}

\item{max.iter}{number of iterations performed for each maximum likelihood step.  Setting this to 1 uses the one-step estimator}
}
\value{
Test of constant curvature based on subsampling
}
\description{
Constant Curvature Test Based on Subsampling approaches.
}
