% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsampleConstantCurvatureTest.R
\name{SubSampleConstantCurvatureTestMultipleThresholds}
\alias{SubSampleConstantCurvatureTestMultipleThresholds}
\title{Compute the Constant Curvature Test using Multiple Thresholds. Used in replicating the simulations.}
\usage{
SubSampleConstantCurvatureTestMultipleThresholds(
  D.hat,
  D.subsample,
  reference.set,
  tri.const.seq,
  verbose = F
)
}
\arguments{
\item{D.hat}{initial estimated distance matrix}

\item{D.subsample}{list of subsampled distance matrices}

\item{reference.set}{reference set of selected midpoints and filtered reference points}

\item{tri.const.seq}{sequence of filtered x.values}

\item{verbose}{print details}
}
\value{
Test of constant curvature using multiple thresholds
}
\description{
Compute the Constant Curvature Test using Multiple Thresholds. Used in replicating the simulations.
}
