% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateCurvature.R
\name{EstimateCurvature}
\alias{EstimateCurvature}
\title{Estimate Curvature for an input Distance Matrix}
\usage{
EstimateCurvature(
  D.hat,
  tri.const = 1.4,
  num.midpoints = 3,
  d.yz.min = 1.5,
  d.yz.max = Inf,
  verbose = F
)
}
\arguments{
\item{D.hat}{EstimateD Distance matrix}

\item{tri.const}{Triangle Constant to use for selecting good x points}

\item{num.midpoints}{Number of midpoint sets to use}

\item{d.yz.min}{Minimum distance of y z pairs to consider}

\item{d.yz.max}{Maximum distance of y z pairs to consider}

\item{verbose}{Additional plotting details}
}
\value{
Return estimated values of curvature
}
\description{
Estimate Curvature for an input Distance Matrix
}
