% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateCurvature.R
\name{EstimateKappa}
\alias{EstimateKappa}
\title{Find the solution to the implicit estimating function to estimate curvature}
\usage{
EstimateKappa(
  dxy,
  dxz,
  dyz,
  dxm,
  max.iter = 10,
  ee.thresh = 0.001,
  abs.ee.thresh = 10^(-9),
  kappa.thresh = 1e-04,
  min.curvature = -5000,
  init.gridsize = 10000,
  gridsearch = T
)
}
\arguments{
\item{dxy}{Triangle Side Length}

\item{dxz}{Triangle Side Length}

\item{dyz}{Triangle Side Length}

\item{dxm}{Length from point x to the midpoint of y and z}

\item{max.iter}{Maximum number of iterations}

\item{ee.thresh}{Threshold for the ratio of the solution of the estimating function}

\item{abs.ee.thresh}{Threshold for the absolute value of the solution of the estimating function}

\item{kappa.thresh}{Threshold for the gridsearch precision}

\item{min.curvature}{Minimum possible value of curvature}

\item{init.gridsize}{Gridsize to pick for good initialization}

\item{gridsearch}{Whether to use grid-search for estimating curvature as opposed to a Newton Method.}
}
\value{
Estimated Curvature Value
}
\description{
Find the solution to the implicit estimating function to estimate curvature
}
\examples{
EstimateKappa(1,1,1,sqrt(3/4))
}
