% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsampleConstantCurvatureTest.R
\name{SubSampleConstantCurvatureTestMultipleThresholds}
\alias{SubSampleConstantCurvatureTestMultipleThresholds}
\title{Compute the Constant Curvature Test using Multiple Thresholds. Used in replicating the simulations.}
\usage{
SubSampleConstantCurvatureTestMultipleThresholds(
  A,
  clique.set,
  D.subsample,
  tri.const.seq,
  J = 3,
  verbose = F
)
}
\arguments{
\item{A}{Adjacency matrix}

\item{clique.set}{indices of where the cliques are located in A}

\item{D.subsample}{list of subsampled Distance Matrices}

\item{tri.const.seq}{sequence of filtered x.values}

\item{J}{number of midpoints}

\item{verbose}{print details}
}
\value{
Test of constant curvature using multiple thresholds
}
\description{
Compute the Constant Curvature Test using Multiple Thresholds. Used in replicating the simulations.
}
